move QueryStringDecoder to core
