+++ http-poja-common/src/main/java/io/micronaut/http/poja/PojaHttpRequest.java
@@ -35,7 +35,7 @@ import io.micronaut.http.body.ByteBody.SplitBackpressureMode;
 import io.micronaut.http.body.CloseableByteBody;
 import io.micronaut.http.codec.MediaTypeCodec;
 import io.micronaut.http.codec.MediaTypeCodecRegistry;
-import io.micronaut.http.poja.util.QueryStringDecoder;
+import io.micronaut.servlet.http.utils.QueryStringDecoder;
 import io.micronaut.servlet.http.ServletExchange;
 import io.micronaut.servlet.http.ServletHttpRequest;
